test: GameSituation 테스트코드
